(feat: bindings) Revised logic to reduce the code complexity and share more logic between wasm and universal bindings
