+++ src/bindings/mod.rs
@@ -1,7 +1,5 @@
-mod logic;
+/// Universal bindings for all platforms
+pub mod universal;
 /// wasm prover and verifier
 #[cfg(all(target_arch = "wasm32", target_os = "unknown"))]
 pub mod wasm;
-/// Universal bindings for all platforms
-#[cfg(target_os = "ios")]
-pub mod universal;
